requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==24.2.0
discord==2.3.2
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
load-dotenv==0.1.0
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
wg-easy-api-wrapper==1.0.8
yarl==1.18.3
